Add assembly resolve hook for robust assembly dependencies lookup
Currently tool is unable to fetch attributes for older .NET versions
and assemblies targeting .NET Core. In this commit we add an
assembly load hook to look for the missing dependencies in the
current .NET Framework distribution. In most cases that helps and
tool is able to fetch assembly attributes.