Support sibling snupkg /symbols.nupkg files
When opening a nupkg file from disk, if its symbols are in a .snupkg or .symbols.snupkg file next to it, it would be flagged as having no symbols. This commit tests for such a file, using it if possible, before returning to the prior behaviour of checking nuget.org for packages identified as coming from there.