(#1480) Fix desktop shortcut in Chocolatey Prior to version 1.0.0 of Chocolatey the Install-DesktopLink helper function could be used to create shortcuts. This function has been deprecated and now removed from Chocolatey as of version 1.0.0
This commit refactors the chocolateyInstall.ps1 script to use the supported
Install-ChocolateyShortcut function to provide the functionality.

Additionally, package parameters have been introduced to allow folks installing
from the chocolatey package to opt-out of having a desktop icon created for them.

This additional functionality, including usage, has been documented in the nuspec
file description, so that it is available for people looking at the package on the
Chocolatey Community Repository.